2.x ahead master

+++ Controller/UserController.php
@@ -3,28 +3,38 @@
 namespace KRG\UserBundle\Controller;
 
 use KRG\UserBundle\Entity\UserInterface;
-use KRG\UserBundle\Form\Type\ChangePasswordType;
+use KRG\UserBundle\Form\Type\ConfirmType;
 use KRG\UserBundle\Form\Type\ProfileType;
 use KRG\UserBundle\Manager\UserManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Component\Finder\Exception\AccessDeniedException;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use KRG\UserBundle\Form\Type\ChangePasswordType;
+use KRG\UserBundle\Manager\LoginManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class UserController extends AbstractController
 {
     /** @var UserManagerInterface */
     protected $userManager;
 
+    /** @var LoginManagerInterface */
+    protected $loginManager;
+
     /** @var TranslatorInterface */
     protected $translator;
 
-    public function __construct(UserManagerInterface $userManager, TranslatorInterface $translator)
+    /** @var TokenStorageInterface */
+    protected $tokenStorage;
+
+    public function __construct(UserManagerInterface $userManager, TranslatorInterface $translator, TokenStorageInterface $tokenStorage)
     {
         $this->userManager = $userManager;
         $this->translator = $translator;
+        $this->tokenStorage = $tokenStorage;
     }
 
     /**
@@ -55,9 +65,12 @@ class UserController extends AbstractController
         }
 
         $form = $this
-            ->createForm(ProfileType::class)
-            ->setData($user)
-            ->add('submit', SubmitType::class, ['label' => 'form.submit_profile']);
+            ->createForm(ProfileType::class, $user, [
+                'action' => $this->generateUrl('krg_user_edit'),
+            ])
+            ->add('submit', SubmitType::class, [
+                'label'  => 'form.submit_profile',
+            ]);
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -87,37 +100,60 @@ class UserController extends AbstractController
         }
 
         $form = $this
-            ->createForm(ChangePasswordType::class)
-            ->setData($user)
-            ->add('submit', SubmitType::class, ['label' => 'form.submit_change_password']);
+            ->createForm(ChangePasswordType::class, $user, [
+                'action' => $this->generateUrl('krg_user_change_password'),
+            ])
+            ->add('submit', SubmitType::class, [
+                'label'  => 'form.user.submit_change_password',
+            ]);
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $user = $form->getData();
 
             $this->userManager->updateUser($user, true);
-            $this->addFlash('notice', $this->translator->trans('change_password.flash.success', [], 'KRGUserBundle'));
+            $this->addFlash('success', $this->translator->trans('change_password.flash.success', [], 'KRGUserBundle'));
 
             return $this->redirectToRoute('krg_user_show');
         }
 
-        return $this->render('@KRGUser/security/changePassword.html.twig', [
+        return $this->render('@KRGUser/user/changePassword.html.twig', [
             'form' => $form->createView(),
             'user' => $user,
         ]);
     }
 
     /**
-     * @Route("profile/delete", name="krg_user_delete")
+     * @Route("/profile/delete", name="krg_user_delete")
      */
-    public function deleteAction()
+    public function deleteAction(Request $request)
     {
         $user = $this->getUser();
+        $form = $this
+            ->createFormBuilder(null, [
+                'translation_domain' => 'KRGUserBundle'
+            ])
+            ->add('submit', SubmitType::class, [
+                'label' => 'user_delete.submit'
+            ])
+            ->getForm();
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($user);
-        $entityManager->flush();
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($user);
+            $entityManager->flush();
 
-        return $this->redirectToRoute('krg_user_login');
+            $this->tokenStorage->setToken(null);
+            $request->getSession()->invalidate();
+
+            $this->addFlash('success', $this->translator->trans('user_delete.success', [], 'KRGUserBundle'));
+
+            return $this->redirectToRoute('krg_user_login');
+        }
+
+        return $this->render('@KRGUser/user/delete.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 }
